--- server/src/routes/auth.js
@@ -1,11 +1,0 @@
-// import express from 'express';
-// import { signup ,login} from '../controllers/authController.js'; 
-
-// const router = express.Router();
-
-
-// router.post('/signup', signup);
-// router.post('/login', login);
-
-
-// export default router;
